fix(association): optimize mission loading Fixes #554
