+++ admin/src/scenes/association/list.js
@@ -3,6 +3,7 @@ import { ReactiveBase, MultiDropdownList, DataSearch } from "@appbaseio/reactive
 
 import { useSelector } from "react-redux";
 import styled from "styled-components";
+import { toastr } from "react-redux-toastr";
 
 import api from "../../services/api";
 import { apiURL } from "../../config";
@@ -16,6 +17,8 @@ const FILTERS = ["SEARCH", "REGION", "DEPARTMENT", "DOMAIN"];
 
 export default () => {
   const [structure, setStructure] = useState(null);
+  const [missionsInfo, setMissionsInfo] = useState({});
+
   const user = useSelector((state) => state.Auth.user);
   const getDefaultQuery = () => ({
     query: {
@@ -32,6 +35,54 @@ export default () => {
       },
     },
   });
+
+  const getMissionsInfo = async (association) => {
+    var myHeaders = new Headers();
+    myHeaders.append("Content-Type", "application/json");
+
+    var raw = JSON.stringify({
+      query: {
+        bool: {
+          must: [
+            {
+              multi_match: {
+                query: association._source.id_rna,
+                fields: ["associationRna", "organizationRNA"],
+              },
+            },
+            {
+              match: {
+                deleted: "no",
+              },
+            },
+          ],
+        },
+      },
+    });
+
+    try {
+      const res = await fetch("https://api.api-engagement.beta.gouv.fr/v0/mission/search", {
+        method: "POST",
+        headers: myHeaders,
+        body: raw,
+        redirect: "follow",
+      }).then((response) => response.json());
+
+      if (res.hits.hits.length > 0) {
+        return {
+          [association._id]: {
+            countMissions: res.hits.hits.length,
+            countPlaces: res.hits.hits.reduce((prev, curr) => (curr._source.places ? prev + curr._source.places : 0), 0),
+            missions: res.hits.hits.map((el) => ({ id: el._id, ...el._source })),
+          },
+        };
+      }
+    } catch (err) {
+      console.log(err);
+      toastr.error("Erreur lors de la récupération des missions");
+      return {};
+    }
+  };
 
   return (
     <div>
@@ -103,13 +154,33 @@ export default () => {
                     </StatsResult>
                   );
                 }}
-                onData={({ rawData }) => {
+                onData={async ({ rawData }) => {
                   // if (rawData?.hits?.hits) setStructureIds(rawData.hits.hits.map((e) => e._id));
+                  if (rawData?.hits?.hits) {
+                    rawData.hits.hits.forEach(async (association) => {
+                      const associationMissions = await getMissionsInfo(association);
+                      setMissionsInfo((prevState) => ({ ...prevState, ...associationMissions }));
+                    });
+                  }
                 }}
                 dataField={undefined}
                 sortBy={undefined}
                 render={({ data }) => {
-                  return data.map((hit) => <Association hit={hit} key={hit._id} onClick={() => setStructure(hit)} selected={structure?._id === hit._id} />);
+                  return data.map((hit) => (
+                    <Association
+                      hit={hit}
+                      missionsInfo={
+                        missionsInfo[hit._id] || {
+                          countMissions: 0,
+                          countPlaces: 0,
+                          missions: [],
+                        }
+                      }
+                      key={hit._id}
+                      onClick={() => setStructure(hit)}
+                      selected={structure?._id === hit._id}
+                    />
+                  ));
                 }}
               />
             </ResultWrapper>
